src/requirements.txt
-i https://pypi.org/simple

# Core Dependencies
numpy>=1.26.4
pandas>=2.2.0
pytz>=2024.1
python-dateutil>=2.8.2
typing-extensions>=4.9.0

# Data Processing
pyarrow>=14.0.1
fastparquet>=2024.2.0

# API and HTTP
requests>=2.31.0
urllib3>=2.0.0  # Updated for Python 3.13 compatibility
python-dotenv>=1.0.0
click>=8.1.7

# AWS
boto3>=1.34.0
botocore>=1.34.0

# Testing and Coverage (optional)
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
coverage>=7.4.0

# Development Tools (optional)
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0
tox>=4.12.0

# Type Stubs (optional, development only)
types-python-dateutil>=2.8.19.20
types-pytz>=2024.1.0.20240203
types-boto3>=1.0.2

